+++ daemons/mrpd/mvrp.c
@@ -639,7 +639,7 @@ int mvrp_recv_msg(void)
 
 int
 mvrp_emit_vidvectors(unsigned char *msgbuf, unsigned char *msgbuf_eof,
-		     int *bytes_used, int lva)
+		     size_t *bytes_used, int lva)
 {
 	mrpdu_vectorattrib_t *mrpdu_vectorptr;
 	uint16_t numvalues;
@@ -886,7 +886,7 @@ int mvrp_txpdu(void)
 {
 	unsigned char *msgbuf, *msgbuf_wrptr;
 	int msgbuf_len;
-	int bytes = 0;
+	size_t bytes = 0;
 	eth_hdr_t *eth;
 	mrpdu_t *mrpdu;
 	unsigned char *mrpdu_msg_ptr;
@@ -981,7 +981,7 @@ int mvrp_txpdu(void)
 	} else
 		goto out;
 
-	msgbuf_len = mrpdu_msg_ptr - msgbuf;
+	msgbuf_len = (int)(mrpdu_msg_ptr - msgbuf);
 
 	bytes = mrpd_send(mvrp_socket, msgbuf, msgbuf_len, 0);
 #if LOG_MVRP
